internship category and waitlist controllers configured
